+++ packages/agtree/.eslintrc.cjs
@@ -56,6 +56,7 @@ module.exports = {
         ],
         'no-restricted-syntax': ['error', 'LabeledStatement', 'WithStatement'],
         'no-continue': 'off',
+        'no-new': 'off',
         'jsdoc/require-param-type': 'off',
         'jsdoc/require-returns-type': 'off',
         'jsdoc/tag-lines': [